// src/API.ts
export type projects = {
  __typename: "projects",
  colorEndScreenMain?: string | null,
  colorEndScreenSub?: string | null,
  colorExperienceScreenHint?: string | null,
  colorExperienceScreenMain?: string | null,
  colorExperienceScreenSub?: string | null,
  colorFontMain?: string | null,
  colorFontSub?: string | null,
  colorFotoScreenMain?: string | null,
  colorFotoScreenSub?: string | null,
  colorLoadingScreenMain?: string | null,
  colorLoadingScreenSub?: string | null,
  colorStartScreenMain?: string | null,
  colorStartScreenSub?: string | null,
  createdAt?: number | null,
  fontLoadingHeading?: string | null,
  fontLoadingMain?: string | null,
  id: string,
  imageEndScreenBackground?: string | null,
  imageEndScreenLogo?: string | null,
  imageExperienceScreenAvatar?: string | null,
  imageExperienceScreenLogo?: string | null,
  imageFotoScreenScreenBackground?: string | null,
  imageFotoScreenScreenLogo?: string | null,
  imageLoadingScreenBackground?: string | null,
  imageLoadingScreenLogo?: string | null,
  imageStartScreenBackground?: string | null,
  imageStartScreenLogo?: string | null,
  layoutActiveEndScreenStatus?: boolean | null,
  layoutActiveExperienceScreenFotoFrameStatus?: boolean | null,
  layoutActiveExperienceScreenLogoStatus?: boolean | null,
  layoutActiveFotoScreenStatus?: boolean | null,
  layoutActiveLoadingScreenStatus?: boolean | null,
  layoutActiveStartScreenStatus?: boolean | null,
  linkEndScreenButtonOne?: string | null,
  linkEndScreenButtonThree?: string | null,
  linkEndScreenButtonTwo?: string | null,
  paymentIntentId?: string | null,
  projectExperienceLayout?: string | null,
  projectLink?: string | null,
  projectName?: string | null,
  projectStatus?: string | null,
  projectType?: string | null,
  projectUrl?: string | null,
  templateId?: string | null,
  textEndScreenButtonOne?: string | null,
  textEndScreenButtonThree?: string | null,
  textEndScreenButtonTwo?: string | null,
  textEndScreenStatus?: string | null,
  textExperienceScreenFotoFrameStatus?: string | null,
  textExperienceScreenHint?: string | null,
  textExperienceScreenImprintContent?: string | null,
  textExperienceScreenImprintHeading?: string | null,
  textExperienceScreenInstructionsContentOne?: string | null,
  textExperienceScreenInstructionsContentThree?: string | null,
  textExperienceScreenInstructionsContentTwo?: string | null,
  textExperienceScreenInstructionsHeading?: string | null,
  textExperienceScreenLogoStatus?: string | null,
  textFotoScreenHint?: string | null,
  textFotoScreenStatus?: string | null,
  textLoadingScreenHint?: string | null,
  textLoadingScreenStatus?: string | null,
  textStartScreenButton?: string | null,
  textStartScreenHeading?: string | null,
  textStartScreenStatus?: string | null,
  textStartScreenSubheading?: string | null,
  updatedAt?: number | null,
};

export type CreateUserInput = {
  available_templates?: string | null,
  email: string,
  name?: string | null,
};

export type User = {
  __typename: "User",
  available_templates?: string | null,
  email: string,
  id: string,
  name?: string | null,
};
// ...
export type TableIDFilterInput = {
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
};

export type TableStringFilterInput = {
  beginsWith?: string | null,
  between?: Array< string | null > | null,
  contains?: string | null,
  eq?: string | null,
  ge?: string | null,
  gt?: string | null,
  le?: string | null,
  lt?: string | null,
  ne?: string | null,
  notContains?: string | null,
};

export type projectTemplateConnection = {
  __typename: "projectTemplateConnection",
  items?:  Array<projectTemplate | null > | null,
  nextToken?: string | null,
};

export type TableProjectsFilterInput = {
  colorExperienceHint?: TableStringFilterInput | null,
  colorExperienceMain?: TableStringFilterInput | null,
  colorExperienceSub?: TableStringFilterInput | null,
  colorLoadingSub?: TableStringFilterInput | null,
  colorMain?: TableStringFilterInput | null,
  fontLoadingHeading?: TableStringFilterInput | null,
  fontLoadingMain?: TableStringFilterInput | null,
  id?: TableIDFilterInput | null,
  imageAvatar?: TableStringFilterInput | null,
  imageEndScreen?: TableStringFilterInput | null,
  imageExperienceScreen?: TableStringFilterInput | null,
  imageFotoScreen?: TableStringFilterInput | null,
  imageLoadingScreen?: TableStringFilterInput | null,
  imageStartScreen?: TableStringFilterInput | null,
  projectExperienceLayout?: TableStringFilterInput | null,
  projectLink?: TableStringFilterInput | null,
  projectName?: TableStringFilterInput | null,
  projectStatus?: TableStringFilterInput | null,
  projectType?: TableStringFilterInput | null,
  projectUrl?: TableStringFilterInput | null,
  textExperienceHint?: TableStringFilterInput | null,
  textExperienceImprint?: TableStringFilterInput | null,
  textExperienceInstructions?: TableStringFilterInput | null,
  textLoadingScreenHint?: TableStringFilterInput | null,
  textStartScreenHeading?: TableStringFilterInput | null,
  textStartScreenSubheading?: TableStringFilterInput | null,
  userId?: TableStringFilterInput | null,
};

export type projectsConnection = {
  __typename: "projectsConnection",
  items?:  Array<projects | null > | null,
  nextToken?: string | null,
};
// ...
export type GetProjectsQueryVariables = {
  id: string,
};

export type GetProjectsQuery = {
  getProjects?:  {
    __typename: "projects",
    colorEndScreenMain?: string | null,
    colorEndScreenSub?: string | null,
    colorExperienceScreenHint?: string | null,
    colorExperienceScreenMain?: string | null,
    colorExperienceScreenSub?: string | null,
    colorFontMain?: string | null,
    colorFontSub?: string | null,
    colorFotoScreenMain?: string | null,
    colorFotoScreenSub?: string | null,
    colorLoadingScreenMain?: string | null,
    colorLoadingScreenSub?: string | null,
    colorStartScreenMain?: string | null,
    colorStartScreenSub?: string | null,
    createdAt?: number | null,
    fontLoadingHeading?: string | null,
    fontLoadingMain?: string | null,
    id: string,
    imageEndScreenBackground?: string | null,
    imageEndScreenLogo?: string | null,
    imageExperienceScreenAvatar?: string | null,
    imageExperienceScreenLogo?: string | null,
    imageFotoScreenScreenBackground?: string | null,
    imageFotoScreenScreenLogo?: string | null,
    imageLoadingScreenBackground?: string | null,
    imageLoadingScreenLogo?: string | null,
    imageStartScreenBackground?: string | null,
    imageStartScreenLogo?: string | null,
    layoutActiveEndScreenStatus?: boolean | null,
    layoutActiveExperienceScreenFotoFrameStatus?: boolean | null,
    layoutActiveExperienceScreenLogoStatus?: boolean | null,
    layoutActiveFotoScreenStatus?: boolean | null,
    layoutActiveLoadingScreenStatus?: boolean | null,
    layoutActiveStartScreenStatus?: boolean | null,
    linkEndScreenButtonOne?: string | null,
    linkEndScreenButtonThree?: string | null,
    linkEndScreenButtonTwo?: string | null,
    paymentIntentId?: string | null,
    projectExperienceLayout?: string | null,
    projectLink?: string | null,
    projectName?: string | null,
    projectStatus?: string | null,
    projectType?: string | null,
    projectUrl?: string | null,
    templateId?: string | null,
    textEndScreenButtonOne?: string | null,
    textEndScreenButtonThree?: string | null,
    textEndScreenButtonTwo?: string | null,
    textEndScreenStatus?: string | null,
    textExperienceScreenFotoFrameStatus?: string | null,
    textExperienceScreenHint?: string | null,
    textExperienceScreenImprintContent?: string | null,
    textExperienceScreenImprintHeading?: string | null,
    textExperienceScreenInstructionsContentOne?: string | null,
    textExperienceScreenInstructionsContentThree?: string | null,
    textExperienceScreenInstructionsContentTwo?: string | null,
    textExperienceScreenInstructionsHeading?: string | null,
    textExperienceScreenLogoStatus?: string | null,
    textFotoScreenHint?: string | null,
    textFotoScreenStatus?: string | null,
    textLoadingScreenHint?: string | null,
    textLoadingScreenStatus?: string | null,
    textStartScreenButton?: string | null,
    textStartScreenHeading?: string | null,
    textStartScreenStatus?: string | null,
    textStartScreenSubheading?: string | null,
    updatedAt?: number | null,
  } | null,
};

export type GetUploadURLQueryVariables = {
  assetType: assetType,
  fileType: fileType,
  projectId: string,
};

export type GetUploadURLQuery = {
  getUploadURL?: string | null,
};

export type GetUserQuery = {
  getUser?:  {
    __typename: "User",
    available_templates?: string | null,
    email: string,
    id: string,
    name?: string | null,
  } | null,
};
// ...
export type ListProjectsQueryVariables = {
  filter?: TableProjectsFilterInput | null,
  limit?: number | null,
  nextToken?: string | null,
};

export type ListProjectsQuery = {
  listProjects?:  {
    __typename: "projectsConnection",
    items?:  Array< {
      __typename: "projects",
      colorEndScreenMain?: string | null,
      colorEndScreenSub?: string | null,
      colorExperienceScreenHint?: string | null,
      colorExperienceScreenMain?: string | null,
      colorExperienceScreenSub?: string | null,
      colorFontMain?: string | null,
      colorFontSub?: string | null,
      colorFotoScreenMain?: string | null,
      colorFotoScreenSub?: string | null,
      colorLoadingScreenMain?: string | null,
      colorLoadingScreenSub?: string | null,
      colorStartScreenMain?: string | null,
      colorStartScreenSub?: string | null,
      createdAt?: number | null,
      fontLoadingHeading?: string | null,
      fontLoadingMain?: string | null,
      id: string,
      imageEndScreenBackground?: string | null,
      imageEndScreenLogo?: string | null,
      imageExperienceScreenAvatar?: string | null,
      imageExperienceScreenLogo?: string | null,
      imageFotoScreenScreenBackground?: string | null,
      imageFotoScreenScreenLogo?: string | null,
      imageLoadingScreenBackground?: string | null,
      imageLoadingScreenLogo?: string | null,
      imageStartScreenBackground?: string | null,
      imageStartScreenLogo?: string | null,
      layoutActiveEndScreenStatus?: boolean | null,
      layoutActiveExperienceScreenFotoFrameStatus?: boolean | null,
      layoutActiveExperienceScreenLogoStatus?: boolean | null,
      layoutActiveFotoScreenStatus?: boolean | null,
      layoutActiveLoadingScreenStatus?: boolean | null,
      layoutActiveStartScreenStatus?: boolean | null,
      linkEndScreenButtonOne?: string | null,
      linkEndScreenButtonThree?: string | null,
      linkEndScreenButtonTwo?: string | null,
      paymentIntentId?: string | null,
      projectExperienceLayout?: string | null,
      projectLink?: string | null,
      projectName?: string | null,
      projectStatus?: string | null,
      projectType?: string | null,
      projectUrl?: string | null,
      templateId?: string | null,
      textEndScreenButtonOne?: string | null,
      textEndScreenButtonThree?: string | null,
      textEndScreenButtonTwo?: string | null,
      textEndScreenStatus?: string | null,
      textExperienceScreenFotoFrameStatus?: string | null,
      textExperienceScreenHint?: string | null,
      textExperienceScreenImprintContent?: string | null,
      textExperienceScreenImprintHeading?: string | null,
      textExperienceScreenInstructionsContentOne?: string | null,
      textExperienceScreenInstructionsContentThree?: string | null,
      textExperienceScreenInstructionsContentTwo?: string | null,
      textExperienceScreenInstructionsHeading?: string | null,
      textExperienceScreenLogoStatus?: string | null,
      textFotoScreenHint?: string | null,
      textFotoScreenStatus?: string | null,
      textLoadingScreenHint?: string | null,
      textLoadingScreenStatus?: string | null,
      textStartScreenButton?: string | null,
      textStartScreenHeading?: string | null,
      textStartScreenStatus?: string | null,
      textStartScreenSubheading?: string | null,
      updatedAt?: number | null,
    } | null > | null,
    nextToken?: string | null,
  } | null,
};